1、优化采购模块数量不能为小数点问题2、新增客服巡查巡楼记录功能模块3、新增页面导出功能4、优化记录查询条件新增时间段查询条件5、其他界面性调整

+++ service-store/src/main/java/com/java110/store/bmo/allocation/IAllocationBMO.java
@@ -9,5 +9,5 @@ public interface IAllocationBMO {
      * @param tmpAllocationStorehouseDto
      * @param allocationStock
      */
-    void doToAllocationStorehouse(AllocationStorehouseDto tmpAllocationStorehouseDto, int allocationStock);
+    void doToAllocationStorehouse(AllocationStorehouseDto tmpAllocationStorehouseDto, double allocationStock);
 }
